+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Basic metrics module (GSGP.Metrics) : @FelipeRosa
 
 ## 0.1.0 - 2016-10-06
 ### Added